Refactor file appPaths.js to paths.js
