+++ src/views/HomeView/HomeView.js
@@ -4,7 +4,7 @@ import { List, ListItem, ListItemText } from '@material-ui/core';
 import { injectIntl, intlShape } from 'react-intl';
 import Container from '../../components/Container';
 import SearchBar from '../../components/SearchBar';
-import { generatePath } from '../../../config/appPaths';
+import { generatePath } from '../../../config/paths';
 
 // TODO: Fix close by events and services lists with actual data items once data is accessible
